Designed new API with retrocompatibility
